requirements.txt
numpy==2.3.5
python-dotenv==1.2.1
uvicorn==0.38.0
fastapi==0.123.5
pyspark==3.5.7
requests==2.32.5